+++ deepinv/models/tv.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 
 
-class TV(nn.Module):
+class TVDenoiser(nn.Module):
     r"""
     Proximal operator of the isotropic Total Variation operator.
 
@@ -26,6 +26,15 @@ class TV(nn.Module):
     :param float crit: Convergence criterion. Default: 1e-5.
     :param torch.tensor, None x2: Primary variable. Default: None.
     :param torch.tensor, None u2: Dual variable. Default: None.
+
+    .. note::
+        The regularization term :math:\|Dx\|_{1,2} is implicitly normalized by its Lipschitz constant, i.e.
+        :math:`\sqrt{8}`, see e.g. A. Beck and M. Teboulle, "Fast gradient-based algorithms for constrained total
+        variation image denoising and deblurring problems", IEEE T. on Image Processing. 18(11), 2419-2434, 2009.
+
+    .. warning::
+        For using TV as a prior for Plug and Play algorithms, it is recommended to use the class
+        :class:`~deepinv.optim.prior.TVPrior` instead. In particular, it allows to evaluate TV.
     """
 
     def __init__(
@@ -36,7 +45,7 @@ class TV(nn.Module):
         x2=None,
         u2=None,
     ):
-        super(TV, self).__init__()
+        super(TVDenoiser, self).__init__()
 
         self.verbose = verbose
         self.n_it_max = n_it_max
@@ -46,7 +55,7 @@ class TV(nn.Module):
         self.tau = 0.01  # > 0
 
         self.rho = 1.99  # in 1,2
-        self.sigma = 1 / self.tau / 72
+        self.sigma = 1 / self.tau / 8
 
         self.x2 = x2
         self.u2 = u2
@@ -65,6 +74,13 @@ class TV(nn.Module):
         )
 
     def forward(self, y, ths=None):
+        r"""
+        Computes the proximity operator of the TV norm.
+
+        :param torch.Tensor y: Noisy image.
+        :param float, torch.Tensor ths: Regularization parameter.
+        :return: Denoised image.
+        """
         restart = (
             True
             if (self.restart or self.x2 is None or self.x2.shape != y.shape)
@@ -84,9 +100,9 @@ class TV(nn.Module):
         for _ in range(self.n_it_max):
             x_prev = self.x2.clone()
 
-            x = self.prox_tau_fx(self.x2 - self.tau * nablaT(self.u2), y)
+            x = self.prox_tau_fx(self.x2 - self.tau * self.nabla_adjoint(self.u2), y)
             u = self.prox_sigma_g_conj(
-                self.u2 + self.sigma * nabla(2 * x - self.x2), lambd
+                self.u2 + self.sigma * self.nabla(2 * x - self.x2), lambd
             )
             self.x2 = self.x2 + self.rho * (x - self.x2)
             self.u2 = self.u2 + self.rho * (u - self.u2)
@@ -100,47 +116,32 @@ class TV(nn.Module):
                     print("TV prox reached convergence")
                 break
 
-            if _ % 100 == 0 and self.verbose:
-                primalcost = 0.5 * torch.linalg.norm(
-                    self.x2.flatten() - y.flatten()
-                ) ** 2 + lambd * torch.sum(
-                    torch.sqrt(torch.sum(nabla(self.x2) ** 2, axis=-1))
-                )
-                dualcost = (y**2).sum() / 2 - torch.sum(
-                    (y - nablaT(self.u2)) ** 2
-                ) / 2.0
-                primalcostlowerbound = max(primalcost, dualcost)
-                print("Iter ", _, "primal cost :", primalcost.item())
-
         return self.x2
 
+    @staticmethod
+    def nabla(x):
+        r"""
+        Applies the finite differences operator associated with tensors of the same shape as x.
+        """
+        b, c, h, w = x.shape
+        u = torch.zeros((b, c, h, w, 2), device=x.device).type(x.dtype)
+        u[:, :, :-1, :, 0] = u[:, :, :-1, :, 0] - x[:, :, :-1]
+        u[:, :, :-1, :, 0] = u[:, :, :-1, :, 0] + x[:, :, 1:]
+        u[:, :, :, :-1, 1] = u[:, :, :, :-1, 1] - x[..., :-1]
+        u[:, :, :, :-1, 1] = u[:, :, :, :-1, 1] + x[..., 1:]
+        return u
 
-def nabla(I):
-    b, c, h, w = I.shape
-    G = torch.zeros((b, c, h, w, 2), device=I.device).type(I.dtype)
-    G[:, :, :-1, :, 0] = G[:, :, :-1, :, 0] - I[:, :, :-1]
-    G[:, :, :-1, :, 0] = G[:, :, :-1, :, 0] + I[:, :, 1:]
-    G[:, :, :, :-1, 1] = G[:, :, :, :-1, 1] - I[..., :-1]
-    G[:, :, :, :-1, 1] = G[:, :, :, :-1, 1] + I[..., 1:]
-    return G
-
-
-def nablaT(G):
-    b, c, h, w = G.shape[:-1]
-    I = torch.zeros((b, c, h, w), device=G.device).type(
-        G.dtype
-    )  # note that we just reversed left and right sides of each line to obtain the transposed operator
-    I[:, :, :-1] = I[:, :, :-1] - G[:, :, :-1, :, 0]
-    I[:, :, 1:] = I[:, :, 1:] + G[:, :, :-1, :, 0]
-    I[..., :-1] = I[..., :-1] - G[..., :-1, 1]
-    I[..., 1:] = I[..., 1:] + G[..., :-1, 1]
-    return I
-
-
-# # ADJOINTNESS TEST
-# u = torch.randn((4, 3, 100,100)).type(torch.DoubleTensor)
-# Au = nabla(u)
-# v = torch.randn(*Au.shape).type(Au.dtype)
-# Atv = nablaT(v)
-# e = v.flatten()@Au.flatten()-Atv.flatten()@u.flatten()
-# print('Adjointness test (should be small): ', e)
+    @staticmethod
+    def nabla_adjoint(x):
+        r"""
+        Applies the adjoint of the finite difference operator.
+        """
+        b, c, h, w = x.shape[:-1]
+        u = torch.zeros((b, c, h, w), device=x.device).type(
+            x.dtype
+        )  # note that we just reversed left and right sides of each line to obtain the transposed operator
+        u[:, :, :-1] = u[:, :, :-1] - x[:, :, :-1, :, 0]
+        u[:, :, 1:] = u[:, :, 1:] + x[:, :, :-1, :, 0]
+        u[..., :-1] = u[..., :-1] - x[..., :-1, 1]
+        u[..., 1:] = u[..., 1:] + x[..., :-1, 1]
+        return u