+++ examples/plug-and-play/demo_vanilla_PnP.py
@@ -110,7 +110,7 @@ model = optim_builder(
 # --------------------------------------------------------------------
 #
 # The model returns the output and the metrics computed along the iterations.
-# For cumputing PSNR, the ground truth image ``x_gt`` must be provided.
+# For computing PSNR, the ground truth image ``x_gt`` must be provided.
 
 y = physics(x)
 x_lin = physics.A_adjoint(y)  # linear reconstruction with the adjoint operator