+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ New Features
 ^^^^^^^^^^^^
 - Added r2r loss (:gh:`148` by `Brayan Monroy`_) - 30/01/2024
 - Added scale transform (:gh:`135` by `Jérémy Scanvic`_) - 19/12/2023
+- Added priors for total variation and l12 mixed norm (:gh:`156` by `Nils Laurent`_) - 09/02/2023
 
 
 Fixed